requirements.txt
nornir
netmiko
nornir-utils
pyyaml